Add preliminary phenix 1.21 support for main matchmaps utility
