+++ .idea/misc.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Black">
+    <option name="enabledOnReformat" value="true" />
+    <option name="pathToExecutable" value="$USER_HOME$/miniconda3/bin/black" />
+    <option name="sdkName" value="mmdev" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="mmdev" project-jdk-type="Python SDK" />
+</project>
